fix the checkwinner logic and added a function to show who won the game

+++ script.js
@@ -11,13 +11,14 @@ const playerTwo = PlayerFactory('Player Two', 'O');
 
 const gameBoard = (() => {
     let _board= new Array(9).fill('');
-    let getBoard = () => [..._board]
+    let getBoard = () => [..._board];
+    const win = ''
 
     const placeMarker = (index,marker) => {
          deleteItem = 1;
+         if(_board[index] !== '')return false;
          _board.splice(index,deleteItem,marker);
-        //  console.log(board.getBoard());
-         
+         console.log(_board);
     }
 
     const printBoard = () => {
@@ -39,7 +40,7 @@ const gameBoard = (() => {
           (_board[0] === playerOne.getMarker() && _board[4] === playerOne.getMarker() && _board[8] === playerOne.getMarker()) 
         ) {
             // console.log(playerOne.getName());
-            console.log(`${playerOne.getName()} won`);
+            return 'win';
         }
       
         else if (
@@ -54,16 +55,19 @@ const gameBoard = (() => {
           // diagonal
           (_board[0] === playerTwo.getMarker() && _board[4] === playerTwo.getMarker() && _board[8] === playerTwo.getMarker()) 
         ) {
-            console.log(`${playerTwo.getName()} won`);
+            return 'lose';
         }
-        else if (!_board.includes('')) {
-            console.log('draw')
+        else if (_board.every(cell => cell !== '')) {
+            return 'draw'
         }
-      
+       
       };
-     
+    
+    const resetBoard = () => {
+        _board = new Array(9).fill('');
+    }  
 
-     return {getBoard, placeMarker, printBoard, checkWinner}
+     return {getBoard, placeMarker, printBoard, checkWinner, resetBoard}
 })();
 
 const gameController = (() => {
@@ -80,6 +84,8 @@ const gameController = (() => {
         }
     ]
 
+    const getPlayer = () => [...player]
+
     let activePlayers = player[0];
 
     const switchPlayers = () => {
@@ -94,24 +100,28 @@ const gameController = (() => {
     const getActivePlayer = () => activePlayers;
 
     const playGame = (index) => {
-        console.log(
-            `${getActivePlayer().name} put his marker on cell ${index}`
-        )
+        // console.log(
+        //     `${getActivePlayer().name} put his marker on cell ${index}`
+        // )
         board.placeMarker(index,getActivePlayer().marker)
-        board.checkWinner();    
-    
+        // board.checkWinner();    
         switchPlayers(); 
+    
 
     }
 
-    return { getActivePlayer, playGame}
+    return { getActivePlayer, playGame, getPlayer, checkWinner: board.checkWinner}
 })();
 
 const displayController = (() => {
     const game = gameController;
+    const gameDiv = document.querySelector('.game-control')
     const _boardDiv = document.querySelector('.board');
     const _playerTurn = document.querySelector('.turn');
-
+    const resultContainer = document.querySelector('.alert-container');
+    const result = document.querySelector('.game-result');
+    const resetBtn = document.querySelector('.reset');
+    
    
     const createGrid = () => {
         for(let i = 0; i < 9; i++) {
@@ -124,25 +134,68 @@ const displayController = (() => {
     };
 
     const updateScreen = () => {
-         const activePlayer = game.getActivePlayer().name;   
+        
 
-        _playerTurn.textContent = activePlayer;
+         const activePlayer = game.getActivePlayer().name;
+         const drawMsg = (name,name2) => result.textContent = `${name} and ${name2} is tied.`
+         const winMsg = (name) => result.textContent = `${name} wins`;
+         const loseMsg = (name) => result.textContent = `${name} wins`;
+         const toggleMsgAndBtn = () => {
+            resultContainer.classList.add('display-on');
+            gameDiv.style.display = 'none'
+
+         };
+
+         const isWin = () => {
+            // print the winner of the game;
+            if(game.checkWinner() === 'draw') {
+               toggleMsgAndBtn();
+               return drawMsg(game.getPlayer()[0].name,game.getPlayer()[1].name);
+             }
+             else if(game.checkWinner() === 'win') {
+                toggleMsgAndBtn();
+                return winMsg(game.getPlayer()[0].name);
+                
+             }
+             else if(game.checkWinner() === 'lose') {
+                toggleMsgAndBtn();
+                return loseMsg(game.getPlayer()[1].name);
+
+             }     
+
+             
+    }
+        isWin()
+        
+
+        _playerTurn.textContent = `${activePlayer} turn`;
+
+
     }
 
     const sqrClick = () => {
         const sqr = document.querySelectorAll('.sqr');
         sqr.forEach(cell => cell.addEventListener('click', (e) => {
             const selectedSqr = e.target.getAttribute('index');
+            
+            if(e.target.textContent !=='')return;
             e.target.textContent = `${game.getActivePlayer().marker}`
-            updateScreen();
             game.playGame(selectedSqr);
+            updateScreen();
+
         }));
       
+    }
+
+    const resetBoard = (e) => {
+        resetBtn.addEventListener('click',(e)=>console.log(e.target))
+        
     }
 
     createGrid();
     updateScreen();
     sqrClick();
+    resetBoard();
  
 })();
